Version 1.14.1
o Added '...' to readCodelinkSet() so that it pass additional arguments (e.g. 'flag') to readCodelink(). Thank you to Luca Beltrame for the suggestion and patch.
o Now when readCodelinkSet() reads a list of files and the phenoData is not supplied it will create a default one, instead of failing.

Version: 1.14.0

o Version number bumped for BioC 2.5.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/branches/RELEASE_2_5/madman/Rpacks/codelink@42948 bc3139a8-67e5-0310-9ffc-ced21a209358